proto: move DnsSecError et al into dnssec
